moved fetch for new thoughts to NewThoughtForm
