+++ code/src/components/Thoughts.js
@@ -5,7 +5,6 @@ import './thoughts.css'
 
 export const Thoughts = () => {
   const [thoughts, setThoughts] = useState([])
-  const [newThought, setNewThought] = useState('')
 
   // Fetch happy thoughts from API using GET
   useEffect(() => {
@@ -13,23 +12,6 @@ export const Thoughts = () => {
       .then(res => res.json())
       .then(json => setThoughts(json))
   }, [])
-
-  // Function determining what is happening when form i submitted 
-  const handleThoughtFormSubmit = (event) => {
-    event.preventDefault()
-
-    // Post new messages to API using POST
-    fetch("https://technigo-thoughts.herokuapp.com/", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ message: newThought })
-    })
-      .then((res) => res.json())
-      .then((newThought) => {
-        setThoughts((previousThoughts) => [newThought, ...previousThoughts]) // Adds new thought to array without having to fetch again
-        setNewThought('') // Clears textarea input field
-      })
-  }
 
   const onHeartClicked = (thoughtId) => {
     const heartClicked = thoughts.map(thought => {
@@ -45,9 +27,7 @@ export const Thoughts = () => {
     <div className="thoughts-wrapper">
       <section className="thoughts-container">
         <NewThoughtForm
-          onSubmit={handleThoughtFormSubmit}
-          value={newThought}
-          onChange={(event) => setNewThought(event.target.value)} />
+          setThoughts={setThoughts} />
         <ThoughtsList
           thoughts={thoughts}
           onHeartClicked={onHeartClicked} />
